# Optimize findall performance by preallocating estimated matches